ラベルスタックの中身を32bitで保存復元するようにした

+++ core/iwasm/migration/wasm_dump.c
@@ -239,19 +239,19 @@ _dump_stack(WASMExecEnv *exec_env, struct WASMInterpFrame *frame, struct FILE *f
     for (i = 0; i < ctrl_stack_size; ++i, ++csp) {
         // uint8 *begin_addr;
         addr = get_addr_offset(csp->begin_addr, wasm_get_func_code(frame->function));
-        fwrite(&addr, sizeof(uint64), 1, fp);
+        fwrite(&addr, sizeof(uint32), 1, fp);
 
         // uint8 *target_addr;
         addr = get_addr_offset(csp->target_addr, wasm_get_func_code(frame->function));
-        fwrite(&addr, sizeof(uint64), 1, fp);
+        fwrite(&addr, sizeof(uint32), 1, fp);
 
         // uint32 *frame_sp;
         addr = get_addr_offset(csp->frame_sp, frame->sp_bottom);
-        fwrite(&addr, sizeof(uint64), 1, fp);
+        fwrite(&addr, sizeof(uint32), 1, fp);
 
         // uint32 *frame_tsp;
         addr = get_addr_offset(csp->frame_tsp, frame->tsp_bottom);
-        fwrite(&addr, sizeof(uint64), 1, fp);
+        fwrite(&addr, sizeof(uint32), 1, fp);
         
         // uint32 cell_num;
         fwrite(&csp->cell_num, sizeof(uint32), 1, fp);
