+++ core/iwasm/migration/wasm_migration.h
@@ -25,7 +25,7 @@ get_addr_offset(void* target, void* base)
 }
 
 static void*
-set_addr_offset(void* base, uint64 offset)
+set_addr_offset(void* base, uint32 offset)
 {
     if (offset == -1) return NULL;
     else return base + offset;
